move the common type FECompetitionSummary to feFootballDataPage file
